requirement.txt
flask==1.0.2
flask_sqlalchemy==2.3.2
flask_marshmallow==0.10.0
marshmallow_sqlalchemy==0.16.2